Moved the UI score text to a separate file
